requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
comtypes==1.1.14
dxcam==0.0.5
MouseInfo==0.1.3
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==9.3.0
pip==22.3.1
psutil==5.9.4
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
PyQt5==5.15.7
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
requests==2.28.1
setuptools==60.2.0
urllib3==1.26.12
wheel==0.37.1
